Move various things to new package
