+++ packages/truffle-decoder/lib/interface/index.ts
@@ -2,9 +2,6 @@ import { AstDefinition, Values } from "truffle-decode-utils";
 import { DataPointer } from "../types/pointer";
 import { EvmInfo } from "../types/evm";
 import decode from "../decode";
-import TruffleDecoder from "./contract-decoder";
-import { ContractObject } from "truffle-contract-schema/spec";
-import { Provider } from "web3/providers";
 import { DecoderRequest, GeneratorJunk } from "../types/request";
 
 export { getStorageAllocations, storageSize } from "../allocate/storage";
@@ -12,10 +9,11 @@ export { getCalldataAllocations } from "../allocate/calldata";
 export { getMemoryAllocations } from "../allocate/memory";
 export { readStack } from "../read/stack";
 export { slotAddress } from "../read/storage";
-
-export function forContract(contract: ContractObject, relevantContracts: ContractObject[], provider: Provider, address?: string): TruffleDecoder {
-  return new TruffleDecoder(contract, relevantContracts, provider, address);
-}
+export { StoragePointer, isStoragePointer } from "../types/pointer";
+export { StorageAllocations, StorageMemberAllocations, StorageMemberAllocation } from "../types/allocation";
+export { Slot, isWordsLength } from "../types/storage";
+export { DecoderRequest, isStorageRequest, isCodeRequest } from "../types/request";
+export { EvmInfo } from "../types/evm";
 
 export function* forEvmState(definition: AstDefinition, pointer: DataPointer, info: EvmInfo): IterableIterator<Values.Value | DecoderRequest | GeneratorJunk> {
   return yield* decode(definition, pointer, info);
