+++ packages/truffle-decoder/lib/types/errors.ts
@@ -25,11 +25,3 @@ export class UnknownUserDefinedTypeError extends Error {
     this.typeString = typeString;
   }
 }
-
-export class ContractBeingDecodedHasNoNodeError extends Error {
-  constructor() {
-    const message = "Contract does not appear to have been compiled with Solidity (cannot locate contract node)";
-    super(message);
-    this.name = "ContractBeingDecodedHasNoNodeError";
-  }
-}
